Refactoring from Page to Content View to fix memory leak issue
