+++ Avanade-StudioTV.UWP/Avanade-StudioTV.UWP.csproj
@@ -121,6 +121,7 @@
     <Content Include="Assets\Fonts\FontAwesome4.otf">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
+    <Content Include="TVLogo.PNG" />
     <Content Include="WeatherCloudy.png" />
     <Content Include="WeatherStorm.png" />
     <Content Include="WeatherSunny.png" />
