+++ Avanade-StudioTV/Views/Controls/Converters.cs
@@ -7,6 +7,9 @@ using Xamarin.Forms;
 
 namespace AvanadeStudioTV.Views.Controls
 {
+	/// <summary>
+	/// NOTE: Both of these converters are not used IN CURRENT CODE:
+	/// </summary>
 	public class StringToUpperAndWideSpacingConverter : IValueConverter
 {
 	public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
@@ -25,4 +28,18 @@ namespace AvanadeStudioTV.Views.Controls
 		return null;
 	}
 }
+
+
+	class EqualsParameterContextConverter : IValueConverter
+	{
+		public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+		{
+			return value == ((View)parameter).BindingContext;
+		}
+
+		public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+		{
+			throw new NotImplementedException();
+		}
+	}
 }
